feat(component-update.sh): add script to update compoennts
so far the script needs a tag, a branch and a component
it will then update the component to use the latest commit in the
client branch and give it a tag on the component. The tag can then
be used to run the compoennt with the new version of the client in
odtp.

The script is just in a basic version for now and can be improved
later to take for example a list of components or find a usable
tag on its own